Move rules into rules package and other smaller code cleanups
