+++ internal/app/config/config_test.go
@@ -1,14 +1,14 @@
 package config
 
 import (
-	"github.com/ldebruijn/graphql-protect/internal/business/aliases"
-	"github.com/ldebruijn/graphql-protect/internal/business/batch"
-	"github.com/ldebruijn/graphql-protect/internal/business/block_field_suggestions"
-	"github.com/ldebruijn/graphql-protect/internal/business/enforce_post"
-	"github.com/ldebruijn/graphql-protect/internal/business/max_depth"
 	"github.com/ldebruijn/graphql-protect/internal/business/persisted_operations"
+	"github.com/ldebruijn/graphql-protect/internal/business/rules/aliases"
+	"github.com/ldebruijn/graphql-protect/internal/business/rules/batch"
+	"github.com/ldebruijn/graphql-protect/internal/business/rules/block_field_suggestions"
+	"github.com/ldebruijn/graphql-protect/internal/business/rules/enforce_post"
+	"github.com/ldebruijn/graphql-protect/internal/business/rules/max_depth"
+	"github.com/ldebruijn/graphql-protect/internal/business/rules/tokens"
 	"github.com/ldebruijn/graphql-protect/internal/business/schema"
-	"github.com/ldebruijn/graphql-protect/internal/business/tokens"
 	"github.com/ldebruijn/graphql-protect/internal/http/proxy"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/yaml.v3"
