+++ internal/business/rules/batch/batch_test.go
