+++ internal/http/proxy/proxy.go
@@ -3,7 +3,7 @@ package proxy
 import (
 	"bytes"
 	"encoding/json"
-	"github.com/ldebruijn/graphql-protect/internal/business/block_field_suggestions"
+	"github.com/ldebruijn/graphql-protect/internal/business/rules/block_field_suggestions"
 	"io"
 	"net"
 	"net/http"
